Fixed take_cutting bug, overcovered other bugs.
This fixes a bug whereby when a cutting was taken, the original tree
could end up with an overlapping lft/rgt field where the branch had been
sliced from.

However, whilst working on this bug, I spotted some flawed logic in a
number of other places.. for instance, taking a cutting leaves 'leaf'
nodes that expand to fill the gap. This unfortunately means that the
is_leaf call that identifies leaves by their width can no longer work.

I've included a test for this.